requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
python-dotenv>=0.19.0
python-telegram-bot[job-queue]>=20.0
